Promote input parameter sweeps-per-measurement to double.

+++ include/dca/phys/dca_step/cluster_solver/ss_ct_hyb/ss_ct_hyb_walker.hpp
@@ -314,7 +314,7 @@ void SsCtHybWalker<device_t, parameters_type, MOMS_type>::test_interpolation() {
 
 template <dca::linalg::DeviceType device_t, class parameters_type, class MOMS_type>
 void SsCtHybWalker<device_t, parameters_type, MOMS_type>::do_sweep() {
-  int factor = 1;
+  double factor = 1.;
   if (thermalized)
     factor = parameters.get_sweeps_per_measurement();
